chore: organize core folder
