+++ src/core/auth/strategies/jwt-refresh.strategy.ts
@@ -1,8 +1,8 @@
 import { Injectable, Request } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy, ExtractJwt } from 'passport-jwt';
-import { PayloadToken } from '../../../common/dtos/auth.dto';
 
+import { PayloadToken } from '../../../common/types/auth.type';
 @Injectable()
 export class JwtRefreshStrategy extends PassportStrategy(
   Strategy,
