+++ src/core/auth/strategies/jwt.strategy.ts
@@ -1,7 +1,8 @@
 import { ExtractJwt, Strategy } from 'passport-jwt';
 import { PassportStrategy } from '@nestjs/passport';
 import { Injectable } from '@nestjs/common';
-import { PayloadToken } from '../../../common/dtos/auth.dto';
+
+import { PayloadToken } from '../../../common/types/auth.type';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy, 'jwt') {
