+++ src/core/auth/strategies/local.strategy.ts
@@ -1,7 +1,9 @@
-import { Strategy } from 'passport-local';
 import { PassportStrategy } from '@nestjs/passport';
 import { Injectable, UnauthorizedException } from '@nestjs/common';
-import { AuthService } from '../../services/auth.service';
+
+import { Strategy } from 'passport-local';
+
+import { AuthService } from '../../../core/services/auth.service';
 
 @Injectable()
 export class LocalStrategy extends PassportStrategy(Strategy, 'local') {
