+++ src/core/modules/auth.module.ts
@@ -1,13 +1,14 @@
-import { JwtRefreshStrategy } from './../auth/strategies/jwt-refresh.strategy';
 import { Module } from '@nestjs/common';
-import { AuthService } from '../services/auth.service';
 import { PassportModule } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
-import { UsersModule } from './users.module';
+import { ConfigService } from '@nestjs/config';
+
 import { AuthController } from '../../api/controllers/auth.controller';
+import { JwtRefreshStrategy } from '../auth/strategies/jwt-refresh.strategy';
 import { JwtStrategy } from '../auth/strategies/jwt.strategy';
 import { LocalStrategy } from '../auth/strategies/local.strategy';
-import { ConfigService } from '@nestjs/config';
+import { AuthService } from '../services/auth.service';
+import { UsersModule } from './users.module';
 
 @Module({
   imports: [UsersModule, PassportModule, JwtModule.register({})],
